Live-Streaming-using-OpenCV-Flask-main/requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==9.1.0
protobuf==3.20.0
redis==3.0.1
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
